send api - moved contract calls
