+++ contracts/examples/lottery-esdt/src/lib.rs
@@ -308,7 +308,7 @@ pub trait Lottery {
 		serializer.push_argument_bytes(esdt_token_name.as_slice());
 		serializer.push_argument_bytes(amount.to_bytes_be().as_slice());
 
-		self.async_call(&to, &BigUint::zero(), serializer.as_slice());
+		self.send().async_call(&to, &BigUint::zero(), serializer.as_slice());
 	}
 
 	fn sum_array(&self, array: &[u8]) -> u16 {
