+++ contracts/examples/multisig/src/lib.rs
@@ -493,7 +493,7 @@ pub trait Multisig {
 					arg_buffer.push_raw_arg(arg.as_slice());
 				}
 				let new_address =
-					self.deploy_contract(gas_left, &amount, &code, code_metadata, &arg_buffer);
+					self.send().deploy_contract(gas_left, &amount, &code, code_metadata, &arg_buffer);
 				result.push(new_address.into_boxed_bytes());
 			},
 			Action::SCCall {
@@ -506,7 +506,7 @@ pub trait Multisig {
 				for arg in arguments {
 					call_data.push_argument_bytes(arg.as_slice());
 				}
-				self.async_call(&to, &amount, call_data.as_slice());
+				self.send().async_call(&to, &amount, call_data.as_slice());
 			},
 		}
